backend/requirements.txt
django>=5.0,<6.0
graphene-django>=3.2
django-filter>=24.0
djangorestframework>=3.14
django-cors-headers>=4.3
celery>=5.4
redis>=5.0
psycopg2-binary>=2.9
gunicorn>=23.0
python-dotenv>=1.0
cryptography>=42.0
bcrypt>=4.0
graphql-relay>=3.2
httpx>=0.27
boto3>=1.34
weasyprint>=62.0
clickhouse-connect>=0.8
pyyaml>=6.0
